add test_kl

+++ divtools/diversity/kl.py
@@ -7,7 +7,6 @@ from divtools.model.game_level_2d import GameLevel2D
 def KL_Divergence(p, q):
     p = np.asarray(p, dtype=np.float64)
     q = np.asarray(q, dtype=np.float64)
-    # divide by zero error
     q = np.where(q != 0, q, np.finfo(float).eps)
 
     return np.sum(np.where(p != 0, p * np.log(p / q), 0))
@@ -24,7 +23,6 @@ def find_pattern(level: GameLevel2D, eye):
         for j in range(level_map.shape[1] - (eye - 1)):
             # Extract the 2x2 sub-array
             sub_array = level_map[i:i + eye, j:j + eye]
-
             # Convert the sub-array to a tuple and count it
             patterns[tuple(map(tuple, sub_array))] += 1
 
@@ -34,7 +32,7 @@ def find_pattern(level: GameLevel2D, eye):
 def KL_Divergence_2d(level1: GameLevel2D, level2: GameLevel2D, eye):
     patterns_original = find_pattern(level1, eye)
     patterns_copy = copy.deepcopy(patterns_original)
-    level2_map = np.asarray(level2.map, dtype=np.int32)
+    level2_map = np.asarray(level2.map)
     pattern_key = set(patterns_original.keys())
     for p in patterns_copy:
         patterns_copy[p] = 0
@@ -50,4 +48,3 @@ def KL_Divergence_2d(level1: GameLevel2D, level2: GameLevel2D, eye):
     values2 = [patterns_copy[key] for key in pattern_key]
 
     return KL_Divergence(values1, values2)
-
